% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_keystrokes.R
\name{plot_keystrokes}
\alias{plot_keystrokes}
\title{Plot keystrokes on a keyboard}
\usage{
plot_keystrokes(string, layout = "qwerty")
}
\arguments{
\item{string}{character, a character string to visualize on a keyboard.}

\item{layout}{character, define the keyboard layout. Default is 'qwertz'.}
}
\value{
ggplot2 object
}
\description{
This function takes a character string and fills the keys on a plotted keyboard depending on the appearance of each character.
}
\details{
The provided character string will be split into single upper-case characters and each character then gets matched to a given keyboard layout dataset. Besides the respective character or letter the keyboard layout dataset also includes the positions of each key on the keyboard. According to the number of the appearance of each character the single keys on the keyboard get filled.
}
\examples{
plot_keystrokes(string = "Hello! Isn't R 3.2.5 just great?", layout = "qwerty")
}

